requirements.txt
Babel==1.3
Django==1.8.2
django-appconf==1.0.1
django-compressor==1.5
django-cors-headers==1.0.0
django-extra-views==0.6.4
django-filter==0.9.2
django-filters==0.1.0
django-frontend==1.6.1
django-haystack==2.3.1
django-localflavor==1.1
django-oscar==1.0.2
django-oscar-paypal==0.9.5
django-paypal==0.2.3
django-tables2==1.0.4
django-treebeard==3.0
django-widget-tweaks==1.3
djangorestframework==3.1.1
drf-nested-routers==0.9.0
factory-boy==2.4.1
Markdown==2.6.2
mock==1.0.1
phonenumbers==6.3.0
Pillow==2.8.1
purl==1.1
pycountry==1.10
pytz==2015.2
requests==2.7.0
six==1.9.0
sorl-thumbnail==11.12.1b0
South==1.0.2
Unidecode==0.4.17
wheel==0.24.0
